FEAT: Se realiza un refactor y mejoras del código en general.
1- Se agrega nuevas funcionalidades y endpoints (casos de uso).
2- Actualiza los test ya existentes.
3- Se castea los errores y agregan excepciones en la capa de dominio (lógica).
4- Actualizamos el README.md
5- Mejora de lógica de base de datos que la pasamos de la capa de infrastruture a la capa application, esto mejora la migración y no dependemos de ninguna tecnologia en particular.

+++ src/main/java/com/challange/api/rest/banco/dominio/model/Cliente.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @Builder
 public class Cliente {
 
-    private int idCliente;
+    private Long idCliente;
     private String nombres;
     private String apellidos;
     private String telefono;
